+++ server/src/matchTypes.ts
@@ -1,10 +1,48 @@
-import { MatchType } from "./parserClasses";
 import { Match } from "./parserClasses";
+/**
+ * The different types a match returned by the peggy parser for ISG-CNC files can have.
+ */
+
+export enum MatchType {
+    mainPrg = "mainPrg",
+    localSubPrg = "localSubPrg",
+    mainPrgName = "mainPrgName",
+
+    // prg calls
+    localPrgCall = "localPrgCall",
+    localPrgCallName = "localPrgCallName",
+    localPrgDefinitionName = "localPrgDefinitionName",
+    globalPrgCall = "globalPrgCall",
+    globalPrgCallName = "globalPrgCallName",
+    localCycleCall = "localCycleCall",
+    localCycleCallName = "localCycleCallName",
+    globalCycleCall = "globalCycleCall",
+    globalCycleCallName = "globalCycleCallName",
+    cycleParameter = "cycleParameter",
+    cycleParameterAssignment = "cycleParameterAssignment",
+    cycleParamList = "cycleParamList",
+
+    // other
+    toolCall = "toolCall",
+    controlBlock = "controlBlock",
+    gotoBlocknumber = "gotoBlocknumber",
+    gotoLabel = "gotoLabel",
+    label = "label",
+    multiline = "multiline",
+    trash = "trash",
+    skipBlock = "skipBlock",
+    blockNumber = "blockNumber",
+    blockNumberLabel = "blockNumberLabel",
+    varDeclaration = "varDeclaration",
+    variable = "variable",
+    lineComment = "lineComment",
+    blockComment = "blockComment"
+};
 
 /**
  * Returns the according definition-type to a given match and if the definition has
  * to be searched locally or globally
- * @param match 
+ * @param match the match to get the definition type for 
  * @returns \{ defType: string | null, local: boolean } an object containing the definition type and a boolean indicating if the definition has to be searched locally or globally
  */
 export function getDefType(match: Match): { defType: MatchType | null, local: boolean } {