+++ server/src/parserClasses.ts
@@ -1,5 +1,6 @@
 import * as peggy from "peggy";
 import { WorkDoneProgressServerReporter } from "vscode-languageserver";
+import { MatchType } from "./matchTypes";
 
 /**
  * A position in a text document expressed as zero-based line and character offset.
@@ -39,7 +40,7 @@ export class FileRange {
 export interface ParseResultContent {
     fileTree: Array<any>;
     numberableLinesUnsorted: Set<number>;
-    mainPrg: Match | null;
+    mainPrgLoc: peggy.LocationRange | null;
 }
 
 /**
@@ -66,7 +67,7 @@ export interface Match {
 }
 
 /** Returns if a given object is a Match and so can be converted to such*/
-export function hasMatchProperties(obj: any): boolean {
+export function isMatch(obj: any): obj is Match {
     const exampleMatch = {
         type: null,
         content: null,
@@ -125,40 +126,3 @@ export class IncrementableProgress {
         return this.canceled;
     }
 }
-
-/**
- * The different types a match returned by the peggy parser for ISG-CNC files can have.
- */
-
-export enum MatchType {
-    toolCall = "toolCall",
-    mainPrg = "mainPrg",
-    localSubPrg = "localSubPrg",
-
-    // prg calls
-    localPrgCall = "localPrgCall",
-    localPrgCallName = "localPrgCallName",
-    localPrgDefinitionName = "localPrgDefinitionName",
-    globalPrgCall = "globalPrgCall",
-    globalPrgCallName = "globalPrgCallName",
-    localCycleCall = "localCycleCall",
-    localCycleCallName = "localCycleCallName",
-    globalCycleCall = "globalCycleCall",
-    globalCycleCallName = "globalCycleCallName",
-    cycleParameter = "cycleParameter",
-    cycleParameterAssignment = "cycleParameterAssignment",
-    cycleParamList = "cycleParamList",
-    
-    controlBlock = "controlBlock",
-    gotoBlocknumber = "gotoBlocknumber",
-    gotoLabel = "gotoLabel",
-    label = "label",
-    multiline = "multiline",
-    trash = "trash",
-    skipBlock = "skipBlock",
-    blockNumber = "blockNumber",
-    blockNumberLabel = "blockNumberLabel",
-    varDeclaration = "varDeclaration",
-    variable = "variable",
-    comment = "comment"
-};