+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Types of changes:
 ## [Unreleased]
 
 ### Added
+- Hover information for calls of self-defined subrograms/cycles based on program comments above their definition [Issue #155](https://github.com/isg-stuttgart/vscode-isg-cnc/issues/155)
 
 ### Fixed